2 files modified

+++ pages/blitz.js
@@ -36,9 +36,9 @@ let lookTable=()=>{
     let arr = [];
       tds.forEach((t,x)=>{
         //console.log(x,t.textContent);
-        if (x===0)      arr.push(parseInt(t.dataset.time));
+        if (x===0)      arr.push(parseFloat(t.dataset.time));
         else if (x===8) arr.push(t.textContent);
-        else            arr.push(parseInt(t.textContent));
+        else            arr.push(parseFloat(t.textContent));
       });
        obj[rowid]=arr;
   });
